+++ packages/gafl-webapp-service/src/__tests__/server.spec.js
@@ -1,35 +1,7 @@
-'use strict'
+import { start, stop, server } from '../test-utils.js'
 
-/**
- * Not a real mock hapi but hapi with no catbox-redis
- */
-
-import { createServer, init, server } from '../server.js'
-import CatboxMemory from '@hapi/catbox-memory'
-
-createServer({
-  cache: [
-    {
-      provider: {
-        constructor: CatboxMemory
-      }
-    }
-  ]
-})
-
-beforeEach(async done => {
-  server.events.on('start', () => {
-    done()
-  })
-  await init()
-})
-
-afterEach(done => {
-  server.events.on('stop', () => {
-    done()
-  })
-  server.stop()
-})
+beforeAll(d => start(d))
+afterAll(d => stop(d))
 
 test('Server is alive', () => {
   expect(server.info).toBeTruthy()
